+++ Common/FrontierCompression.h
@@ -3,7 +3,7 @@
 #include "Buffer.h"
 
 namespace FrontierCompression {
-    int BitMapSize(int count, uint32_t* indexes);
+    size_t BitMapSize(int count, uint32_t* indexes);
     int EncodeBitMap(int count, uint32_t* indexes, uint8_t* buffer, int buffer_capacity);
     bool IsBitMap(uint8_t* buffer);
     int DecodeBitMap(int& size, uint8_t* buffer, uint32_t* indexes, int values_capacity);
